requirements.txt
llama-index
llama-index-llms-openai
llama-index-readers-file
pypdf
python-dotenv
pillow
pdfminer.six